Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -13,7 +13,7 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i = 0;
+	unsigned int i = 0;
 
 	if (array == NULL)
 	{
@@ -21,7 +21,7 @@ int linear_search(int *array, size_t size, int value)
 	}
 	while (i < size)
 	{
-		printf("Value checked array[%lu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%u] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
